+++ app/src/types.ts
@@ -1,0 +1,10 @@
+import { Player } from '../../src'
+
+import { Asset } from './Asset'
+
+export type PageProps = {
+  player: Player
+  asset?: Asset
+  autoload?: boolean
+}
+