+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Removed `InterstitialPlayer` component. Support for HLS interstitials is
   now integrated into the standard player.
 
+## New Features
+
+* Add ad-related UI and styling to existing components.
+
 # 0.8.1 (Beta)
 
 ## Fixes